+++ models/Thought.js
@@ -1,0 +1,22 @@
+import mongoose from "mongoose";
+
+const Thought = mongoose.model("Thought", {
+  message: {
+    type: String,
+    required: true,
+    minlength: 5,
+    maxlength: 140,
+  },
+
+  heart: {
+    type: Number,
+    default: 0,
+  },
+
+  createdAt: {
+    type: Date,
+    default: Date.now,
+  },
+});
+
+export default Thought;
